requirements.txt
fastapi
opencv-python
Pillow